% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ACCURAATHEID}
\alias{ACCURAATHEID}
\title{Constante voor de Accuraatheid}
\format{An object of class \code{integer} of length 1.}
\usage{
ACCURAATHEID
}
\description{
Constante voor de Accuraatheid
}
\keyword{constant}
\keyword{dqf}
