% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQScoringFramework.R
\name{addScoreToDQFramework}
\alias{addScoreToDQFramework}
\title{A Function to transpose a data based on some parameters}
\usage{
addScoreToDQFramework(categorie, waarde, weging)
}
\arguments{
\item{categorie}{categorie}

\item{waarde}{waarde 1-5}

\item{weging}{weging van het item}
}
\description{
This function blablablabla
}
\examples{
addScoreToDQFramework(category, waarde, weging)
}
\keyword{dcast}
\keyword{melt}
\keyword{transposed}
