% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{COMPLEETHEID}
\alias{COMPLEETHEID}
\title{Constante voor de compleetheid}
\format{An object of class \code{integer} of length 1.}
\usage{
COMPLEETHEID
}
\description{
Constante voor de compleetheid
}
\keyword{constant}
\keyword{dqf}
