% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CONSISTENTIE}
\alias{CONSISTENTIE}
\title{Constante voor de Consistentie}
\format{An object of class \code{integer} of length 1.}
\usage{
CONSISTENTIE
}
\description{
Constante voor de Consistentie
}
\keyword{constant}
\keyword{dqf}
