% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dumpRDS}
\alias{dumpRDS}
\title{A Function to save the rds to a location whithin the datafiles-location}
\usage{
dumpRDS(object, naam, ...)
}
\arguments{
\item{object}{naam van het object dat bewaard moet worden}

\item{naam}{naam van het bestand}

\item{...}{overige parameters}
}
\description{
This function om de exacte locatie te achterhalen
}
\keyword{file}
\keyword{location}
\keyword{rds}
