% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLocationNaam}
\alias{getLocationNaam}
\title{A Function to get the filename on the local pc.}
\usage{
getLocationNaam(naam, checkExists = TRUE)
}
\arguments{
\item{naam}{naam van het bestand}

\item{checkExists}{is er controle nodig om te kijken of het bestand bestaat}
}
\description{
This function om de exacte locatie te achterhalen
}
\keyword{file}
\keyword{local}
\keyword{location}
