% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeDQScoringFramework.R
\name{initializeDQScoringFramework}
\alias{initializeDQScoringFramework}
\title{A Function to transpose a data based on some parameters}
\usage{
initializeDQScoringFramework()
}
\description{
This function blablablabla
}
\examples{
initializeDQScoringFramework()
}
\keyword{DQ}
\keyword{Scoring}
\keyword{init}
