% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQScoringFramework.R
\name{plotDQ}
\alias{plotDQ}
\title{Een functie die het dataframe resultaat in een dashboard toont}
\usage{
plotDQ()
}
\description{
Een functie die het dataframe resultaat in een super cool dashboard toont
}
\keyword{DQFramework}
\keyword{dashboard}
