% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readCSV}
\alias{readCSV}
\title{A Function to read a csv-file from a location whithin the work-location}
\usage{
readCSV(naam, ...)
}
\arguments{
\item{naam}{naam van het bestand}

\item{...}{overige parameters}
}
\description{
This function om de exacte locatie te achterhalen
}
\keyword{csv}
\keyword{file}
\keyword{location}
