% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readRDSdd}
\alias{readRDSdd}
\title{A Function to read a rds-file from a location whithin the work-location}
\usage{
readRDSdd(naam, ...)
}
\arguments{
\item{naam}{naam van het bestand}

\item{...}{overige parameters}
}
\description{
This function om de exacte locatie te achterhalen
}
\keyword{file}
\keyword{location}
\keyword{rds}
