% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setWorkdir}
\alias{setWorkdir}
\title{A Function to specify the data location}
\usage{
setWorkdir(directory)
}
\arguments{
\item{directory}{naam van de basisdirectory}
}
\description{
This function om de exacte locatie te achterhalen
}
\keyword{file}
\keyword{local}
\keyword{location}
