% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preparation.R
\name{startPreparation}
\alias{startPreparation}
\title{A Function to initialize dataframes to the datadiggers project}
\usage{
startPreparation(workdir, rebuild = FALSE,
  dataframesToGlobalEnvironment = FALSE)
}
\arguments{
\item{workdir}{directory where the sourcefiles are located}

\item{rebuild}{indication of dataframes have to be rebuild when not available}

\item{dataframesToGlobalEnvironment}{infication of dataframes to global enironment}
}
\description{
This function
}
\examples{
startPreparation( "C:/files", TRUE, FALSE)
}
\keyword{eda}
\keyword{import}
\keyword{prepare}
