% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{UNICITEIT}
\alias{UNICITEIT}
\title{Constante voor de Uniciteit}
\format{An object of class \code{integer} of length 1.}
\usage{
UNICITEIT
}
\description{
Constante voor de Uniciteit
}
\keyword{constant}
\keyword{dqf}
