% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{VALIDITEIT}
\alias{VALIDITEIT}
\title{Constante voor de validiteit}
\format{An object of class \code{integer} of length 1.}
\usage{
VALIDITEIT
}
\description{
Constante voor de validiteit
}
\keyword{constant}
\keyword{dqf}
