% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLocationNaam}
\alias{getLocationNaam}
\title{Bepalen bestandsnaam en controleren of het bestand bestaat}
\usage{
getLocationNaam(naam, checkExists = TRUE)
}
\arguments{
\item{naam}{naam van het bestand}

\item{checkExists}{is er controle nodig om te kijken of het bestand bestaat}
}
\description{
Bepalen bestandsnaam en controleren of het bestand bestaat
}
\keyword{file}
\keyword{local}
\keyword{location}
