% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQScoringFramework.R
\name{initializeDQScoringFramework}
\alias{initializeDQScoringFramework}
\title{A Function to initialize the data-quality-framework}
\usage{
initializeDQScoringFramework()
}
\description{
This function blablablabla
}
\examples{
initializeDQScoringFramework()
}
\keyword{DQ}
\keyword{Scoring}
\keyword{init}
