% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQScoringFramework.R
\name{initializeDQScoringFramework}
\alias{initializeDQScoringFramework}
\title{Initialisatie van het data-quality-framework}
\usage{
initializeDQScoringFramework()
}
\description{
Deze funtie initaliseert het framework om datakwaliteit te scoren.
}
\examples{
initializeDQScoringFramework()
}
\keyword{DQ}
\keyword{Scoring}
\keyword{init}
