% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQScoringFramework.R
\name{plotDQ}
\alias{plotDQ}
\title{A Function display the results of the DQF}
\usage{
plotDQ()
}
\description{
This function plots the marvelous DataQuality-dashboard
}
\keyword{DQFramework}
\keyword{dashboard}
