% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesmining.R
\name{procesMining.init}
\alias{procesMining.init}
\title{DataPreparation for procesMining with Bupar}
\usage{
procesMining.init(pStartTijd, pEindTijd, pThresholds)
}
\arguments{
\item{pStartTijd}{starttijd, waarbinnen een order moet vallen om een cnet te bepalen}

\item{pEindTijd}{Eindtijd, waarbinnen een order moet vallen om een cnet te bepalen}

\item{pThresholds}{collectie van thresholds waarvoor een causal net wordt gemaakt}
}
\description{
DataPreparation for procesMining with Bupar
}
