% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preparation.R
\name{startPreparation}
\alias{startPreparation}
\title{Een functie die de ruwe bestanden verwerkt tot basis dataframes}
\usage{
startPreparation(workdir, rebuild = FALSE,
  dataframesToGlobalEnvironment = FALSE)
}
\arguments{
\item{workdir}{Directory waar de data-bestanden zich bevinden}

\item{rebuild}{Indicatie of alles opnieuw gebouwd dient te worden}

\item{dataframesToGlobalEnvironment}{indicatie of de dataframes in het global environment beschikbaar gemaakt moeten worden}
}
\description{
Deze functie maakt van de ruwe bestanden dataframes.
Indien mogelijk wordt hier gebruik gemaakt van reeds bestaande dumps
Indien aangegeven worden de bestaande dumps juist niet gebruikt maar alles opnieuw bepaald.
}
\keyword{eda}
\keyword{import}
\keyword{prepare}
