% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposeWorkflowDataFrame.R
\name{transposeWorkflowDataFrame}
\alias{transposeWorkflowDataFrame}
\title{A Function to transpose a data based on some parameters}
\usage{
transposeWorkflowDataFrame(data, measureVars, dcastFormula)
}
\arguments{
\item{the}{dataframe that has to be transposed}

\item{a}{list of measured values}

\item{the}{formula that is used to transpose via dcast}
}
\description{
This function blablablabla
}
\examples{
transposeWorkflowDataFrame( workflowDF, c("measure1", "measure2"),)
}
\keyword{dcast}
\keyword{melt}
\keyword{transposed}
