% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposeWorkflowDataFrame.R
\name{transposeWorkflowDataFrame}
\alias{transposeWorkflowDataFrame}
\title{A Function to transpose a data based on some parameters}
\usage{
transposeWorkflowDataFrame(data, measureVars, dcastFormula)
}
\arguments{
\item{data}{het dataframe dat als input dient}

\item{measureVars}{lijst met variabelen die worden gemeten}

\item{dcastFormula}{de formule die gebruikt wordt voor het transponeren van het dataframe}
}
\description{
Deze functie kantel het ingegeven dataframe en bewaard hierbij de ingegeven meetwaarden.
}
\keyword{dcast}
\keyword{melt}
\keyword{transposed}
